[chore] Page을 컴포넌트로 분리
