+++ src/components/App.jsx
@@ -1,36 +1,6 @@
-function NumButton({ onClick, children }) {
-  return (
-    <button onClick={() => onClick(children)} type='button'>{children}</button>
-  )
-}
-
-function Buttons({ onClick }) {
-  const numbers = [1, 2, 3, 4, 5];
-  return (
-    <div>
-      {nums.map((i) => (
-        <NumButton key={i} onClick={onClick}>{i}</NumButton>
-      ))}
-    </div>
-
-  )
-}
-
-function MainClick({ onClick, count }) {
-  return (
-    <button onClick={() => onClick(1)} >Click me ({count})</button>
-  )
-}
-
-
-function Page({ onClick, count }) {
-  return (
-    <div>
-      <MainClick onClick={onClick} count={count} />
-      <Buttons onClick={onClick} />
-    </div>
-  )
-}
+import React from 'react'
+import { useState } from 'react';
+import Page from './Page.jsx'
 
 function App() {
   const [state, setState] = useState({
@@ -48,3 +18,6 @@ function App() {
     <Page onClick={handleClick} count={count} />
   )
 }
+
+export default App;
+
