+++ src/index.jsx
@@ -1,6 +1,5 @@
-import React, { useState } from 'react';
+import React from 'react'
 import ReactDOM from 'react-dom';
 import App from './components/App.jsx'
-
 
 ReactDOM.render(<App />, document.getElementById('app'))
